(cleanup) organized auth components into own folder
